Separate the internal API data types from the public ones.
