+++ workload/api/data-internal.go
@@ -6,13 +6,8 @@ package api
 // TODO(ericsnow) Eliminate the params import if possible.
 
 import (
-	"github.com/juju/errors"
-
 	"github.com/juju/juju/apiserver/params"
 )
-
-// BulkFailure indicates that at least one arg failed.
-var BulkFailure = errors.Errorf("at least one bulk arg has an error")
 
 // TrackArgs are the arguments for the Track endpoint.
 type TrackArgs struct {
